Initial functional release (MVP) (#36)
* Add nginx.conf as API gateway (#31)

* Implement protection (#32)

* Implement CORS and XSS

* Enable secure cookies only in production

* Added environment in config.py [UserService]

* Added Circuit Breaker to PaymentService

* Delete get_registration_by_id in registration_router.py

* Init frontend commit

* Add vibe-coded frontend and fixed bugs on backend (#33)

* Complete MailTrap  implementation and fix bugs (#34)

* Implement the MailTrap external API

* Implement function for random generating bib_number and delete bib_number and payment_status from RegistrationBase in schema.py

* Solved some bugs

* Added some cols in PaymentService

* Ref/bugs (#35)

* Implement the MailTrap external API

* Implement function for random generating bib_number and delete bib_number and payment_status from RegistrationBase in schema.py

* Solved some bugs

* Added some cols in PaymentService

* Fixed bug with paymentstatus update

+++ Services/RaceService/app/services/race_service.py
@@ -1,7 +1,7 @@
 from sqlalchemy.ext.asyncio import AsyncSession
 from fastapi import HTTPException, status
 from app.api.schema import RaceCreate, RaceResponse, RaceUpdate
-from app.db.repositories import race_repository
+from app.db.repositories import race_repository, registration_repository
 
 
 # Service functions for Race
@@ -10,6 +10,12 @@ async def get_race_by_id(db: AsyncSession, race_id: int):
     if not race:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Race not found")
     return RaceResponse.model_validate(race)
+
+async def get_races(db: AsyncSession):
+    races = await race_repository.get_races(db)
+    if not races:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Races not found")    
+    return [RaceResponse.model_validate(race) for race in races]
 
 async def create_race(db: AsyncSession, data: RaceCreate, organiser_id: int):
     race = await race_repository.create_race(db, data, organiser_id)
@@ -30,4 +36,6 @@ async def delete_race(db: AsyncSession, race_id: int, organiser_id: int):
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Race not found")
     if race.organiser_id != organiser_id:
         raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to delete this race")
+    if await registration_repository.get_registration_count_by_race(db,race_id)>0:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="You can't do that, race has participants!")
     return await race_repository.delete_race(db, race_id)
